feat: divide trace data source into trace and api trace (#421)
* feat: divide trace data source into trace and api trace

* fix: remove traceId from trace to api-trace

* test: update

* fix: update detail widget model to use Trace or ApiTrace data source model

* fix: make prettier
