+++ projects/distributed-tracing/src/public-api.ts
@@ -96,6 +96,7 @@ export { traceDetailDashboard } from './pages/trace-detail/trace-detail.dashboar
 export { TraceDetailPageComponent } from './pages/trace-detail/trace-detail.page.component';
 
 // Datasources
+export * from './shared/dashboard/widgets/trace-detail/data/trace-detail-data-source.model';
 export * from './shared/dashboard/widgets/span-detail/data/span-detail-data-source.model';
 export * from './shared/dashboard/widgets/trace-detail/data/api-trace-detail-data-source.model';
