+++ projects/distributed-tracing/src/shared/dashboard/widgets/trace-detail/data/trace-detail-data-source.model.test.ts
@@ -8,7 +8,7 @@ import {
 import { ObservedGraphQlRequest } from '../../../data/graphql/graphql-query-event.service';
 import { ApiTraceDetailDataSourceModel } from './api-trace-detail-data-source.model';
 
-describe('Trace detail data source model', () => {
+describe('API Trace detail data source model', () => {
   const testTimeRange = { startTime: new Date(1568907645141), endTime: new Date(1568911245141) };
   let model!: ApiTraceDetailDataSourceModel;
   let emittedQueries: unknown;
@@ -78,8 +78,8 @@ describe('Trace detail data source model', () => {
         timeRange: expect.objectContaining({ from: testTimeRange.startTime, to: testTimeRange.endTime }),
         traceProperties: expect.arrayContaining([
           expect.objectContaining({ name: 'tags' }),
-          expect.objectContaining({ name: 'traceId' }),
-          expect.objectContaining({ name: 'statusCode' })
+          expect.objectContaining({ name: 'statusCode' }),
+          expect.not.objectContaining({ name: 'traceId' })
         ])
       })
     );
